+++ irRecv.cpp
@@ -1,6 +1,11 @@
 #include "IRremote.h"
 #include "IRremoteInt.h"
 
+#ifdef IR_TIMER_USE_ESP32
+hw_timer_t *timer;
+void IRTimer(); // defined in IRremote.cpp
+#endif
+
 //+=============================================================================
 // Decodes the received IR message
 // Returns 0 if no data ready, 1 if data ready.
@@ -122,6 +127,17 @@ IRrecv::IRrecv (int recvpin, int blinkpin)
 //
 void  IRrecv::enableIRIn ( )
 {
+// Interrupt Service Routine - Fires every 50uS
+#ifdef ESP32
+	// ESP32 has a proper API to setup timers, no weird chip macros needed
+	// simply call the readable API versions :)
+	// 3 timers, choose #1, 80 divider nanosecond precision, 1 to count up
+	timer = timerBegin(1, 80, 1);
+	timerAttachInterrupt(timer, &IRTimer, 1);
+	// every 50ns, autoreload = true
+	timerAlarmWrite(timer, 50, true);
+	timerAlarmEnable(timer);
+#else
 	cli();
 	// Setup pulse clock timer interrupt
 	// Prescale /8 (16M/8 = 0.5 microseconds per tick)
@@ -135,6 +151,7 @@ void  IRrecv::enableIRIn ( )
 	TIMER_RESET;
 
 	sei();  // enable interrupts
+#endif
 
 	// Initialize state machine variables
 	irparams.rcvstate = STATE_IDLE;